Updated openstack/openstack Project: openstack-infra/system-config  d5d91d3be49c2daed1e7b6d1e1a33f5f4accca02
Add roles assignment for openstackci and openstackzuul user/projects

In the infracloud, the Member role is not created by default.
We created that with a previous change by adding it to the launcher.
Now we associate that role to the openstackci/openstackzuul user/projects,
so those users are members of their corresponding projects.

Change-Id: I9147b253c7f747f435c773932dc4a8aad1189799