Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  02ae9dc8ed2c246f3e54d5cc06d91eca1c76d7f6
Fix confirm() util true state

The confirm() util incorrectly returns true only when the answer is
false. This patch reverses the logic so that it properly returns true
when the user inputs a true statement.

Include test case for delete-all command usage to accept or abort.
Mocking the input builtin ensures that the behaviour resulting from the
user inputting 'y' or 'n' will remain tested for delete-all.

Change-Id: I742804a8efe7da8ca6aa70f20059d02758910deb
Co-Authored-By: Darragh Bailey <dbailey@hpe.com>
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>