Updated openstack/openstack Project: openstack/glance  f58b8593926c51aa4d05867760aef45ba6eefb51
api-ref: add versions history

The 'versions' response contains the status of the Images API versions,
but does not indicate the releases when the statuses went into effect.
This patch adds a Version History to the 'versions' api-ref page.  The
language used (for example, "Bexar changes") is consistent with what's
been adopted for other parts of the images api-ref (see the discussion
on https://review.openstack.org/#/c/356693/ )

This patch is current for Newton:
- v1 deprecation: already merged as commit
  63e6dbb1eb006758fbcf7cae83e1d2eacf46b4ab
- v2 minor version bump: the dependency stated below

Depends-On: I5d1c4380682efa4c15ff0f294f269c800fe6762a

Change-Id: Id920a3284a4be23032cc4a23e04726fab6d24361
Closes-bug: #1621161
Partial-bug: #1618495