Updated openstack/openstack Project: openstack/openstack-manuals  e9f6757e1b2f9e9156aba54e9359fc3d41fad489
[install] ADMIN_PASSWORD --> ADMIN_PASS

The install guide consistently uses *_PASS for its password
replacement labels. ADMIN_PASSWORD was introduced when updating
keystone for bootstrap.

This broke consistency with both the Environment --> Security
section, and also the Create a domain, projects, users, and roles
section that follows keystone set-up.

In order to avoid an onslaught of bug reports from the copy/paste
crowd, this patch changes ADMIN_PASSWORD to ADMIN_PASS in the
keystone setup section, keeping the other sections at the original
ADMIN_PASS.

Change-Id: Idf8a9033f229ce325f5e4eb9da6d1241b9923cd3