Updated openstack/openstack Project: openstack/nova  af78a00311df97416293da5eedada798b7185ee9
Add get_context helper method

The context.get_admin_context is used in places that it's not necessary,
and that's likely because there's no helper method to retrieve a non
admin context. This adds such a helper method and adds a note to
get_admin_context warning that it's not usually the right choice.

Change-Id: I2e6a2efa4bcdf3f8688897972a6cf8a5af3f90d6