Updated openstack/openstack Project: openstack/nova  2b6a2767bc2e3224b1e7cbabb9674c89a2f348a6
Disable 'supports_migrate_to_same_host' HyperV driver capability

The Hyper-V driver incorrectly enables the
'supports_migrate_to_same_host' capability.

This capability seems to have been introduced having the VMWare
cluster architecture in mind, but it leads to unintended behavior
in case of the HyperV driver.

For this reason, the Hyper-V CI is failing on an recently introduced
tempest test, which asserts that the host has changed.

This change disables this driver capability.

Change-Id: Ibb4f1d4e40ccc98dc297e463b127772a49207d9a
Closes-Bug: #1629040