Updated openstack/openstack Project: openstack/requirements  bedd17add9a493ca639dc08a9125cb79dc9db830
Use ConfigParser instead of SafeConfigParser

in Python 3 The SafeConfigParser class has been renamed
to ConfigParser in Python 3.2 [1]. The alias SafeConfigParser
maybe removed in future versions of Python 3. Use ConfigParser
instead of SafeConfigParser in Python 3.We're switching from
SafeConfigParser to ConfigParser because we're we're not explicitly
using the 'Safe' functionality.So it make the code potentially neater
if SafeConfigParser is even de-aliased in python3.

[1] http://bugs.python.org/issue10627

Change-Id: I932ced300a93040144e89c7f4acc95480b32749d