Updated openstack/openstack Project: openstack/python-novaclient  749fee334604d0c06dc398c38cb19c51dedb8214
Replace requests mocking with requests-mock

The novaclient tests are broken with the master (unreleased) branch of
requests. This is because most of the mocking that novaclient does with
requests in insufficient.

Requests-mock is for exactly this purpose and is already widely used in
novaclient so replace using mock with requests-mock where appropriate.

Closes-Bug: #1630157
Change-Id: Ie91daeb5f0c241eb83b77c04ac1b12e431d6671c