Updated openstack/openstack Project: openstack-infra/project-config  eba480d52e63aad8b914266bf01c9df1a12133bf
Add non-voting python3 jobs to requirements

The requirements code does not work under python3  Add non-voting jobs
for python34 and python35 until we can get python3 support on all
branches and then switch to voting.

I chose not to use the python34-jobs template as that defaults to
running on both the check and gate pipeline, running non-voting jobs on
the gate pipeline seemed like a waste of resources.

Change-Id: I850758403cb8b30447e5adf5124da422319a98e3