Updated openstack/openstack Project: openstack/horizon  7f47b1baecf9542840cec0bcfe68ad523d522634
Move browser support wiki page into the docs

The browser support page should be part of the docs. This change
moves the content of the wiki page into its own file that is accessible
from the index. References to the wiki page have consequently been
removed from the FAQ.

Change-Id: I340203771cb2b0dfa534ad3f8926e599a727e1e9
Partial-Bug: #1625514