Updated openstack/openstack Project: openstack/nova  434c0a31ac6b357f0622ff5950b4aef25a9d4eaf
refresh instances_path when shared storage used

When doing Live migration with shared storage, it happens erratically,
that the check for the shared storage test_file fails. Because the shared
volume is under heavy IO (many instances on many compute nodes) the client
does not immediately sees the new content of the folder. This delay
could take up to 30s.
This can be fixed if the client is forced to refresh the directories
content, which can be achieved by 'touch' on the directory. Doing so,
the test_file is visibile instantly, within ms.
The patch adds a 'touch' on instances_path in check_shared_storage_test_file,
before checking the existence of the file.

Change-Id: I16be39142278517f43e6eca3441a56cbc9561113
Closes-Bug: #1617299