Updated openstack/openstack Project: openstack-infra/devstack-gate  c6430b68e093f0d5645576d5f96bda97d1888f50
Explicitly set overlay bridge ints as up

Notice that nova net multinode jobs fail due to a lack of connectivity
from tempest to cirros VMs. On further inspections br_pub is DOWN on
this host. Setting it up after the fact seems to make it happy so add an
explicit UP state transition to our setup.

Change-Id: I09d3ebdc7459dccf22aa4f75aa24777715f7ffad