Updated openstack/openstack Project: openstack-infra/elastic-recheck  8e5656cd2d70edb979d046de1f56bbe2202039da
Enable more reasonable logging

This commit adds logging to actually figure out what is going on when
running the uncategorized_fails command. As part of this the log format
is changed slightly to make it a bit more clear. It also adds a new CLI
option, --verbose, to enable the debug log level.

Co-Authored-By: Matthew Treinish <mtreinish@kortar.org>

Change-Id: I6ad3b4e77b15de1e31899510850f2852f301c543