Updated openstack/openstack Project: openstack/ironic  3a2c06587656431e44b5670e90b67b19281aa279
Fix iLO drivers to not clear local_gb if its not detected

In certain cases, due to hardware limitation disk data cannot be
retrieved out-of-band. Currently iLO drivers overwrite the
'local_gb' property of node with '0' during inspection even when
it fails to detect it.
This patch fixes this behavior to ensure that the operator provided
value for 'local_gb' are not overwritten if inspection fails to
detect disk size.

Co-Authored-By: Shivanand Tendulker <stendulker@gmail.com>
Closes-bug: 1602269
Change-Id: Idad3e3fa547022da6213b048cd16c91faae3ff14