Updated openstack/openstack Project: openstack-infra/devstack-gate  1b861477ea0fa60b80227b17baa832ed97b06e4e
Copy ssl certs for multinode

This updates the version of ansible because the older version had a bug
that broke the synchronize module for us. As a side effect of this we
can no longer rely on ansible nooping if an inventory group is empty.
This is now an error so now we check if the subnodes list is empty
before running commands against them.

Change-Id: I66b17561cac62b9bcedd67a724153612744aead6