Updated openstack/openstack Project: openstack/nova  9c89e07d17b5eb441682e3b8fad8b270f37f7015
Add some extra logging around external event handling

While debugging a failed event delivery bug, I realized that our
logging of the event path leaves some ambiguity which makes it
difficult to follow. This adds the instance host to the existing
debug message on the API side, which helps to identify that events
are being sent to the right (or wrong) host. Further, which logs
a warning in the case where a compute host receives an instance
message when no thread was waiting for it.

Change-Id: I9a1461493a19750fe5dc11869cfa33d2e6fd3c6f