Updated openstack/openstack Project: openstack/cinder  7d4fb9fee9eab39672d4714c8690a24e32e96590
Sheepdog: fix clone failure

In the case that Cinder and Glance use same Sheepdog cluster as
their backend, a image is clonable without data copying.

When image A will be cloned to volume B, the Sheepdog's internal
mecanism is following:

1. Make a snapshot C of image A using dog command.
2. Clone snapshot C to Volume B using qemu-img command.

The name of snapshot C is [image A's UUID]-temp-snapshot
So it sometimes starts with a number.
But qemu-img cannot accept such a name.

Change-Id: I20bf1dca058fef31b9f244a4fcc3800b62eadf15
Closes-Bug: #1625530