Updated openstack/openstack Project: openstack/ironic  b5d980522575cfe6e397c1aa9b6514f7d5dc8d57
Update configuration reference link to latest draft

Currently, the configuration reference points to Mitaka. This is wrong
as Newton was just released. Additionally, there is also a different
page published from master.

This commit links to the "draft" page from master. This will direct
readers to the right version of the configuration reference
documentation.

Change-Id: I677bfe0757dd99b63c5204b36e5fe68878eeb1f0