Updated openstack/openstack Project: openstack/glance_store  6903ae45e756f0a2e65f811d6b8b7ad045a78800
Do not require entry-point dependencies in tests

In our tests, we load our entry-points using plain pkg_resources and in
doing so were requiring our dependencies to be checked for each plugin.
This was causing test errors on Python 3.5 running on some environments.

By using logic to load the entry-points as modern versions of setuptools
expect us to and defaulting requirement verification to False, we should
avoid these testing errors.

Change-Id: Id58861f72d399758178e0f11f617d2f178eddc4a
Closes-bug: #1630392