Updated openstack/openstack Project: openstack-infra/elastic-recheck  c432984085c046acd3acbb36a46bfee344d3d2e4
Skip hits that don't have data

The e-r graph and uncategorized fails jobs are both
currently failing because we're getting back hits
with empty data for the 'timestamp' attribute because
the value is too large, e.g.:

"[FIELDDATA] Data too large, data for [@timestamp] would "
"be larger than limit of [22469214208/20.9gb]]"

To workaround this for now we check to see if the hit
item list has anything it it before returning the
data for the facet results.

We should follow this up by logging errors on hits that
have bad data and we should remove those indexes.

Change-Id: Icf19af6580632ef52a55d3fb4bed3bced140024a
Closes-Bug: #1630355