Updated openstack/openstack Project: openstack-dev/devstack  91d8a38e16d5dea09df13aa8063a00e31b42efae
Allow default IPv6 route device names to have dots

When dots are used with sysctl, they are reinterpreted as slashes.
Route devices can have dots in their names, so when they are used in a
sysctl command that also uses dots, its dot will be replaced with a
slash, causing an error.

Change-Id: Ie32126a3aa8d646568d7d37ec4874419b9658935
Closes-Bug: #1627770