Updated openstack/openstack Project: openstack/python-keystoneclient  eeb734c245b221af08fa8c316314217e1fb83d8c
Use fixtures from keystoneauth

As keystoneclient and other services rely more on keystoneauth we should
assume that keystoneauth is our base auth library, not keystoneclient
and start to default to the objects provided from there. This will make
it easier to remove these objects when the time comes.

The easiest thing to move is the keystoneclient fixtures in favor of the
keystoneauth fixtures.

Change-Id: I5a784795536cec6c7ca5eead3f13b1e7a6e45346