Updated openstack/openstack Project: openstack/python-keystoneclient  6f1a3da7caf4da103081c2901d0c464a27f11758
Use AUTH_INTERFACE object from keystoneauth

As keystoneclient and other services rely more on keystoneauth we should
assume that keystoneauth is our base auth library, not keystoneclient
and start to default to the objects provided from there. This will make
it easier to remove these objects when the time comes.

Use the AUTH_INTERFACE special object from keystoneauth in most places.
This uses it everywhere that is actually session independant. For
example it is not changed within the keystoneclient auth plugins
themselves as they are directly compatible with keystoneauth.

Change-Id: Ibc1224fca98c852106feb78c624b0b2f22b3a19d