Updated openstack/openstack Project: openstack-infra/tripleo-ci  d895610b5de102bdbdaa627cd7f36a94986ba05e
Increase disk_allocation_ratio to 4

With a max of 60 envs, we're starting to hit scheduling errors due
to lack of disk space on some of the compute nodes.  In reality,
none of the compute nodes are using more than 61% of their disk
and most are under 50%, so a 33% increase in overcommit should be
safe enough.

We may also want to increase the scheduler retries to help with
this problem.  Part of the issue is that most of the compute nodes
have sufficient disk available, but if we happen to get unlucky and
pick three in a row that don't then the instance fails.  More
scheduler retries would help with that.

Change-Id: Ifb2db1ddafd183aa4c9584b406e3b47bf7b0c5a9