Updated openstack/openstack Project: openstack/diskimage-builder  d0dbe276bd0294c093324af865be7ecf12122630
Create (md5|sha256) checksum files for images

In shade, we use both md5 and sha256 checksums to help validate the
integrity of an image. Rather then having nodepool do this each time
for every time, have diskimage-builder create these files when we
build the image.

We've added a flag (disabled by default) to toggle this functionality.

Change-Id: I5815ba69b7d477f1e91dc8ec0c69c86168770964
Signed-off-by: Paul Belanger <pabelanger@redhat.com>