Updated openstack/openstack Project: openstack/heat  55f0daa4cce39135efbff1c0bf924f3aaade4843
Avoid double init of classes with __new__

When __new__ returns an instance of the class (or a subclass), Python will
call __init__ on it afterwards. So if we return an initialised object of a
subclass, we end up calling both __new__ and __init__ twice. Avoid this by
calling object.__new__ and passing the desired subclass instead in
ResourceInfo and Parameter. Equivalent changes for Resource and Template
have been in place for some time.

Change-Id: I4c3f0bc9c1a39b9f0b964ccf1c6c638f86b3753e