Updated openstack/openstack Project: openstack/tempest  ebfc2bbc4ab12717ab4790e631ffd9da77c93be9
Re-use common methods in ServerRescueNegativeTestJSON

This re-uses the common volume create/attach methods from the
base compute API test class.

Note that this changes the behavior of the test in that the volume
API client is used rather than the proxy volumes extension via the
compute API, which is deprecated anyway.

Also, it's a bit less clear now but the volume detach cleanup
order in the test_rescued_vm_detach_volume test is maintained.

Change-Id: I7db083c8fb945e4b315e807461c0dd6a879ffa10