Updated openstack/openstack Project: openstack/tempest  0235f98f143df20dbd17b658f5f9c35f02099411
Adding a new option CONF.network.dvr_extra_resources

This patch is intended to introduce a new config option. It is called
'dvr_extra_resources'. If the option is set to True, extra resources
will be provisioned to bind the router to the L3 agent. Extra resources
need to be provisioned in the Liberty release or older, and are not
required since the Mitaka release.

We will not encounter the bug 1590049 by setting the option to False.

Change-Id: Ib1a6007bbd65182743e1f1ed5819134d748e8d92