Updated openstack/openstack Project: openstack/ironic  de234a41a0685780d1574747fd8b634b64843ec6
Set SUBNETPOOL_PREFIX_V4 to FIXED_RANGE

The default for SUBNETPOOL_PREFIX_V4 in devstack causes external
connectivity to be mostly broken in osic-cloud1 nodes. IPA relies on
external activity to build images, and osic-cloud1 is pretty prevalent
for running CI jobs, therefore this breaks IPA CI.

Setting this to the same range as FIXED_RANGE should eliminate this, or
at least reduce the probability of it happening from 1:(size of /8) to
1:(size of /24).

This should be reverted when the actual problem is fixed in devstack.

Related-Bug: #1629133
Change-Id: I8f0e5405f3e2d6e35418c73f610ac6b779dd75e5