Updated openstack/openstack Project: openstack-dev/bashate  dc419beb2dc2161c3ac753ab32c913a058e26dec
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I512878af85887d8cb7e8b241797ff611a35972c1