Updated openstack/openstack Project: openstack-dev/hacking  6be124c025019c04e31dabe58448a69bf21c76fa
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I52264fdacde1a2e45d390afb681824b70b23f552