Updated openstack/openstack Project: openstack/designate  24efb6b59b812334296d552cbfbb92b9afa4232b
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I1cd8173e606d65417da72ed49bd5192b4f411d14