Updated openstack/openstack Project: openstack/ironic  e18aec628d1f04bc1a41aacd949aa6ec0db14a90
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I5686b21e1bd8e7ebb3a740c41f5b8204a3f396dc