Updated openstack/openstack Project: openstack/neutron-fwaas  13168c003e01912a948796b1f48217bf5c5ce4b3
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Ib60ba3d2159de9869f30cf6087efb688bd762b76