Updated openstack/openstack Project: openstack/neutron-vpnaas  ed5b17add9268f2f31657906ded31710885125c7
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Iba692c8f37bd0d07c175c145c8019e0ffe6179d3