Updated openstack/openstack Project: openstack/openstackdocstheme  f9b3e8bf5b0e309d3998a03f8fa0d57fc473131b
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I8eac4be92199c7ba98badc1dac68be3e71140d40