Updated openstack/openstack Project: openstack/openstack-doc-tools  699e82d2d269bb570e5f7c564887cfc1b357f671
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Iace1593dff20324b86c7245c2217d51d77d9a465