Updated openstack/openstack Project: openstack/oslo.context  b60ab0fb3f0c951b14f5deab283c450694bb5e1a
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I7994e3f3de21f57242b8a0d5af86fe08d2c6b34e