Updated openstack/openstack Project: openstack/oslo.log  2204e0d537ce5156b18611532e5910cfea3c63ac
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I59db2bf536c766b4708530d41804acfbd4269483