Updated openstack/openstack Project: openstack/python-cinderclient  26b4f987d0b22442a923a672efd540df4816b429
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Idf43509a391fa3a71705bf90ded6fa692b78b60b