Updated openstack/openstack Project: openstack/python-ironicclient  c3e28cd88cb13456d0c3dbdd50b944a2df09a402
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I4bb93f3b53cf9a1dc700fcb7bb6e89a295a47548