Updated openstack/openstack Project: openstack/python-neutronclient  0391cf7e481a07b058ea5c5ff1878ea222d64294
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I5b1b979ada5989eef490de13bf11faa49dee2d1d