Updated openstack/openstack Project: openstack/python-novaclient  4c1dab99e3b1e3815d9dc3e646f8133e5cf443d5
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Ic73a5d161244396297c9d4f72ae4c586e7f81cf8