Updated openstack/openstack Project: openstack/python-saharaclient  a781b0434bdebcf1fc673e915e76275eefc60009
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Id24407f62d885f1e8deabe10e467c327a80b08a9