Updated openstack/openstack Project: openstack/sahara-dashboard  38f5ccccbbe13470a47909aedfddbdcd31dc674b
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Ia6549c2d2ada112e4bc4d23aaeaef4bad33c7276