Updated openstack/openstack Project: openstack/nova  290c31beaacdc6aa4c5af0fb3ba30872b2708d8c
Make releasenotes reminder detect added and untracked notes

When 'tox -ereleasenotes' is run, the releasenotes are built from
the committed notes. We have a script that displays a reminder when
it detects uncommitted notes. It doesn't, however, detect newly
added notes or untracked notes.

This adds detection of added and untracked notes for displaying
the reminder message.

Change-Id: I5164f0089eae4a95bc3e68cd3b515ef23556b227