Updated openstack/openstack Project: openstack/cinder  e885cd99edfdb43e2da16adad1c00b712adf5393
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Ie5e108e73a33a8b2ff2e30f06d39afde1487e72f